roles/mellanox_ofed/requirements.test.txt
ansible==4.7.0
ansible-lint[yamllint]
molecule[lint]
molecule-vagrant
python-vagrant
pytest-testinfra
